# Further optimisations on median_filter3d