This reduces the GPU time of the `median_filter3d_cupy` function from 5.8ms to 3.2ms (for the performance test with kernel size 3 (radius=1). 

The block sizes and register usage has been optimised to achieve a higher occupancy, and some small cupy operations are now performed on the host (calls to `cp.ceil` were running on GPU).

Note that for larger kernel sizes, the local array explodes in size and the registers per thread becomes very large. This limits GPU occupancy and slows down the kernel significantly. However, without dramatically changing the kernel structure and increasing memory a lot, this cannot be avoided. For example, moving from kernel size 3 to kernel size 5 triples the runtime To 10ms, and increasing to 7 brings the runtime to 2000ms. Anything more is prohibitive.

Also, we cannot reduce the memory requirement by working in-place, as this would create a race condition (outputs written by some threads affects inputs of other threads).